Move common functionality for tests to utils class
