+++ tests/lib/ocpp/common/test_database_migration_files.hpp
@@ -1,0 +1,23 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2020 - 2024 Pionix GmbH and Contributors to EVerest
+
+#pragma once
+
+#include "database_testing_utils.hpp"
+#include <ocpp/common/database/database_schema_updater.hpp>
+
+class DatabaseMigrationFilesTest : public DatabaseTestingUtils,
+                                   public ::testing::WithParamInterface<std::tuple<std::filesystem::path, uint32_t>> {
+
+protected:
+    const std::filesystem::path migration_files_path;
+    const uint32_t max_version;
+
+public:
+    DatabaseMigrationFilesTest() :
+        DatabaseTestingUtils(),
+        migration_files_path(std::get<std::filesystem::path>(GetParam())),
+        max_version(std::get<uint32_t>(GetParam())) {
+        EXPECT_EQ(this->database->open_connection(), true);
+    }
+};
